Add callbacks in the memory testing class
Properly testing whether atomic operations are used correctly is hard.
To do this for the page table walker, introduce callbacks that can be
used to modify the memory after a memory operation.

The intended usecase is to "swap" page table contents when A/D bit
updates failed and test whether this still results in a correct
translation.